# api/openapi.yaml
openapi: 3.0.3
info:
  title: Kaji Challenge API
  version: 0.3.0
servers:
  - url: http://localhost:8080
security:
  - cookieAuth: []
paths:
...
  /v1/me/nickname:
    patch:
      operationId: patchMeNickname
      summary: Update current user nickname
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateNicknameRequest'
      responses:
        '200':
          description: Nickname updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UpdateNicknameResponse'
  /v1/teams/invites:
    post:
      operationId: postTeamInvite
      summary: Create invite code
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateInviteRequest'
      responses:
        '201':
          description: Invite created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InviteCodeResponse'
...
  /v1/teams/current:
    patch:
      operationId: patchTeamCurrent
      summary: Update current team name
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCurrentTeamRequest'
      responses:
        '200':
          description: Team updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TeamInfoResponse'
...
  /v1/teams/join:
    post:
      operationId: postTeamJoin
      summary: Join team by invite code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JoinTeamRequest'
      responses:
        '200':
          description: Joined
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JoinTeamResponse'
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: kaji_session
  schemas:
...
    CreateInviteRequest:
      type: object
      properties:
        expiresInHours:
          type: integer
          minimum: 1
          maximum: 720
          default: 72
    InviteCodeResponse:
      type: object
      required: [code, teamId, expiresAt]
      properties:
        code:
          type: string
        teamId:
          type: string
        expiresAt:
          type: string
          format: date-time
    JoinTeamRequest:
      type: object
      required: [code]
      properties:
        code:
          type: string
          minLength: 6
          maxLength: 64
    JoinTeamResponse:
      type: object
      required: [teamId]
      properties:
        teamId:
          type: string
    UpdateCurrentTeamRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 50
    TeamInfoResponse:
      type: object
      required: [teamId, name]
      properties:
        teamId:
          type: string
        name:
          type: string
...
    UpdateNicknameRequest:
      type: object
      required: [nickname]
      properties:
        nickname:
          type: string
          minLength: 1
          maxLength: 30
    UpdateNicknameResponse:
      type: object
      required: [nickname, effectiveName]
      properties:
        nickname:
          type: string
        effectiveName:
          type: string
...